requirements_prod.txt
# ============================================================
# requirements_prod.txt — Servidor summerconceito.com.br
# Versão enxuta: remove pacotes desnecessários em produção
# (OneDrive, Google Drive, ferramentas de dev, psycopg2)
# ============================================================

# Framework Web
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
a2wsgi==1.10.7

# Banco de Dados (SQLite em produção / PostgreSQL opcional)
sqlalchemy==2.0.25
alembic==1.13.1

# Autenticação e Segurança
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Geração de PDF
reportlab==3.6.13
python-barcode==0.15.1
qrcode[pil]==7.4.2
Pillow==9.5.0

# Validação e Serialização
pydantic==1.10.13
email-validator==2.1.0

# Utilidades
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
watchdog==3.0.0

# Logging
loguru==0.7.2
